backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
openai>=1.68.0
python-dotenv==1.0.1
pydantic==2.10.3
python-multipart==0.0.20
